requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=2.4.1
torchvision>=0.19.1
tqdm>=4.41.0
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
imutils~=0.5.4
python-tsp
Flask==3.0.3
Flask-Cors==4.0.1
ultralytics==8.2.92
